requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-bootstrap5>=24.3,<25.0
python-dotenv>=1.0,<2.0
drf-spectacular>=0.28,<1.0
